requirements.txt
Flask>=3.1.0
Flask-SocketIO>=5.3.0
Flask-Cors>=4.0.0
Flask-RESTful>=0.3.10
Flask-Testing>=0.8.1
PyYAML>=5.4.1
pytest>=6.2.5
pytest-flask>=1.3.0
